style: Move OrderedMap to its own file.
It's not very tied to custom properties, so it's worth to have it split IMO.

Only two changes that aren't pure code-move:

 * Added a keys() method to allow reading the `index`.
 * Made `remove()` public.
